# Optimise sanitize with string builder